osd/OSD: Simplifying 'r' usage
The r variable was created unassigned and lately assigned to -1.
Even if r is always defined before its usage, sounds like initializing it to
what seems a default value (-1) from the beginning is safer for later code
inclusion.

This patch also remove the r redefinition to an int as it was already an it.
As there is no clear willing of masking the initial r variable, this patch
remove this useless redefinition of it.

Signed-off-by: Erwan Velu <erwan@redhat.com>